requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
langgraph
python-dotenv
fastapi
uvicorn
duckduckgo-search